+++ dcb-hub-admin-ui/src/pages/serviceInfo/catalogMetricsByHostLms/index.tsx
@@ -12,8 +12,9 @@ import Alert from "@components/Alert/Alert";
 import Link from "@components/Link/Link";
 import Loading from "@components/Loading/Loading";
 import { useRouter } from "next/router";
+import { ProcessState } from "@models/ProcessState";
 
-const BibRecordCountByHostLms: NextPage = () => {
+const CatalogMetricsByHostLms: NextPage = () => {
 	const { publicRuntimeConfig } = getConfig();
 	const { data } = useSession();
 	const { t } = useTranslation();
@@ -30,7 +31,7 @@ const BibRecordCountByHostLms: NextPage = () => {
 		const fetchRecords = async () => {
 			try {
 				const response = await axios.get<any[]>(
-					`${publicRuntimeConfig.DCB_API_BASE}/admin/recordCounts`,
+					`${publicRuntimeConfig.DCB_API_BASE}/hostlmss/importIngestDetails`,
 					{
 						headers: { Authorization: `Bearer ${data?.accessToken}` },
 					},
@@ -58,8 +59,81 @@ const BibRecordCountByHostLms: NextPage = () => {
 	});
 
 	const columns: GridColDef[] = [
-		{ field: "sourceSystemName", headerName: "Source system name", flex: 1 },
-		{ field: "recordCount", headerName: "Bib record count", flex: 1 },
+		{
+			field: "name",
+			headerName: t("catalogMetricsByHostLms.source_system"),
+			flex: 0.75,
+		},
+		{
+			field: "ingestEnabled",
+			headerName: t("catalogMetricsByHostLms.harvest_enabled"),
+			flex: 0.5,
+		},
+		{
+			field: "sourceRecordCount",
+			headerName: t("catalogMetricsByHostLms.harvested"),
+			flex: 0.5,
+		},
+		{
+			field: "awaiting",
+			headerName: t("catalogMetricsByHostLms.await"),
+			flex: 0.5,
+			valueGetter: (value, row) => {
+				const states: ProcessState[] = row?.processStates || [];
+				const failure = states.find(
+					(state: ProcessState) => state.value === "PROCESSING_REQUIRED",
+				);
+				return failure ? failure.count : 0;
+			},
+		},
+		{
+			field: "failed",
+			headerName: t("catalogMetricsByHostLms.failed"),
+			flex: 0.5,
+			valueGetter: (value, row) => {
+				const states: ProcessState[] = row?.processStates || [];
+				const failure = states.find(
+					(state: ProcessState) => state.value === "FAILURE",
+				);
+				return failure ? failure.count : 0;
+			},
+		},
+		{
+			field: "ingested",
+			headerName: t("catalogMetricsByHostLms.ingested"),
+			flex: 0.5,
+			valueGetter: (value, row) => {
+				const states: ProcessState[] = row?.processStates || [];
+				const failure = states.find(
+					(state: ProcessState) => state.value === "SUCCESS",
+				);
+				return failure ? failure.count : 0;
+			},
+		},
+		{
+			field: "bibRecordCount",
+			headerName: t("nav.bibs"),
+			flex: 0.5,
+		},
+		{
+			field: "difference",
+			headerName: t("catalogMetricsByHostLms.difference"),
+			flex: 0.75,
+			valueGetter: (value, row) => {
+				const difference: number = row.sourceRecordCount - row.bibRecordCount;
+				return difference;
+			},
+		},
+		{
+			field: "id",
+			headerName: t("catalogMetricsByHostLms.source_system_id"),
+			flex: 1,
+		},
+		{
+			field: "checkPointId",
+			headerName: t("catalogMetricsByHostLms.checkpoint_id"),
+			flex: 1,
+		},
 	];
 
 	if (loading || status === "loading") {
@@ -69,8 +143,8 @@ const BibRecordCountByHostLms: NextPage = () => {
 					title={t("ui.info.loading.document", {
 						//make the first character lowercase
 						document_type:
-							t("bibRecordCountByHostLms.name").charAt(0).toLowerCase() +
-							t("bibRecordCountByHostLms.name").slice(1),
+							t("catalogMetricsByHostLms.name").charAt(0).toLowerCase() +
+							t("catalogMetricsByHostLms.name").slice(1),
 					})}
 					subtitle={t("ui.info.wait")}
 				/>
@@ -78,7 +152,7 @@ const BibRecordCountByHostLms: NextPage = () => {
 		);
 	}
 	return error ? (
-		<AdminLayout title={t("nav.serviceInfo.bibRecordCountByHostLms")}>
+		<AdminLayout title={t("nav.serviceInfo.catalogMetricsByHostLms")}>
 			<Alert
 				severityType="error"
 				alertText={
@@ -86,7 +160,7 @@ const BibRecordCountByHostLms: NextPage = () => {
 						i18nKey={"common.error_loading"}
 						t={t}
 						values={{
-							page_title: t("nav.serviceInfo.bibRecordCountByHostLms"),
+							page_title: t("nav.serviceInfo.catalogMetricsByHostLms"),
 						}}
 						components={{
 							linkComponent: (
@@ -94,7 +168,7 @@ const BibRecordCountByHostLms: NextPage = () => {
 									onClick={handleReload}
 									title={t("ui.action.reload")}
 									key={t("ui.action.reload")}
-									href="bibRecordCountByHostLms"
+									href="catalogMetricsByHostLms"
 								/>
 							),
 						}}
@@ -104,16 +178,25 @@ const BibRecordCountByHostLms: NextPage = () => {
 			/>
 		</AdminLayout>
 	) : (
-		<AdminLayout title={t("nav.serviceInfo.bibRecordCountByHostLms")}>
+		<AdminLayout
+			title={t("nav.serviceInfo.catalogMetricsByHostLms")}
+			docLink="https://openlibraryfoundation.atlassian.net/wiki/x/GgAnyg"
+			subtitle={t("reference.catalog_build")}
+		>
 			<ClientDataGrid
 				columns={columns}
 				data={records}
-				type="bibRecordCountByHostLMS"
+				type="catalogMetricsByHostLms"
 				// We don't want click-through on this grid.
 				selectable={false}
-				noDataTitle={t("bibRecordCountByHostLms.no_results")}
+				noDataTitle={t("catalogMetricsByHostLms.no_results")}
 				// This is how to set the default sort order
-				sortModel={[{ field: "sourceSystemName", sort: "asc" }]}
+				sortModel={[{ field: "name", sort: "asc" }]}
+				columnVisibilityModel={{
+					id: false,
+					checkPointId: false,
+					sourceSystemId: false,
+				}}
 				operationDataType="BibRecord"
 			/>
 		</AdminLayout>
@@ -137,4 +220,4 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 	};
 };
 
-export default BibRecordCountByHostLms;
+export default CatalogMetricsByHostLms;